Added Working Task Querying
The task pages query properly, but they take a while to load. While it loads, the app shows the basic burger king layout. I would suggest adding a spinner in its replacement. editTask.java is not used in this program currently as for some reason it does not work, but I have commented out the code that would start its activity as well as added comments in regards to what I feel like is happening. I will add the updated layout in a separate commit on the same branch as this one (TestedTasksWithMaster)